BUMCHEO-692 change int to float request params order swap

+++ Mcsg.Story.Api/Requests/Stories/StoryChapterOrderSwapR.cs
@@ -2,6 +2,6 @@
 
 public class StoryChapterOrderSwapR
 {
-    public int Order1 { get; set; }
-    public int Order2 { get; set; }
+    public float Order1 { get; set; }
+    public float Order2 { get; set; }
 }
